fix: address code review — equip destination, fish lock, breath type
- bot.py + js_bot.py: forward destination param through equip() call chain
- bot.py: add _fish_lock to prevent concurrent fish() signal theft
- events.py: BreathEvent.oxygen_level typed as int (matches mineflayer)

+++ src/minethon/models/events.py
@@ -141,7 +141,7 @@ class HealthChangedEvent:
 class BreathEvent:
     """Bot oxygen level changed."""
 
-    oxygen_level: float
+    oxygen_level: int
 
 
 @dataclass(frozen=True, slots=True)
